bot.py
import os
import time
from solana.rpc.api import Client
from solana.publickey import PublicKey
from solana.keypair import Keypair
from solana.transaction import Transaction
from spl.token.constants import TOKEN_PROGRAM_ID
from spl.token.instructions import TransferParams, transfer
from solana.rpc.types import TxOpts
import requests

# Configuración del bot
SOLANA_RPC_URL = "https://api.mainnet-beta.solana.com"  # Cambiar a Devnet para pruebas
WALLET_PRIVATE_KEY = os.environ.get("WALLET_PRIVATE_KEY")  # Llave privada de Solflare
BOT_BALANCE_THRESHOLD = 100  # Monto máximo para invertir (en SOL)

# Cliente de Solana
client = Client(SOLANA_RPC_URL)

# Cargar la cuenta del bot
def load_wallet(private_key):
    return Keypair.from_secret_key(bytes(int(x) for x in private_key.split(",")))

wallet = load_wallet(WALLET_PRIVATE_KEY)

# Obtener balance de SOL
def get_balance(public_key):
    balance = client.get_balance(PublicKey(public_key))["result"]["value"]
    return balance / 1e9  # Convertir lamports a SOL

# Buscar tokens nuevos en la blockchain
def find_new_tokens():
    print("Buscando nuevos tokens...")
    # Aquí, haríamos una consulta a un servicio como Serum o Raydium
    response = requests.get("https://api.solscan.io/token?type=new&limit=10")
    tokens = response.json()["data"]
    return tokens

# Analizar un token para determinar si es prometedor
def analyze_token(token_data):
    # Implementar lógica para analizar un token
    # Ejemplo: checar volumen, liquidez, holders únicos
    volume = token_data.get("volume", 0)
    liquidity = token_data.get("liquidity", 0)
    holders = token_data.get("holders", 0)
    if volume > 1000 and liquidity > 500 and holders < 1000:  # Ejemplo de filtros
        return True
    return False

# Comprar un token
def buy_token(token_address, amount):
    print(f"Comprando token {token_address} por {amount} SOL...")
    # Implementar transacción de compra en Raydium o Serum
    pass

# Vender un token
def sell_token(token_address, amount):
    print(f"Vendiendo token {token_address} por {amount} SOL...")
    # Implementar transacción de venta en Raydium o Serum
    pass

# Ciclo principal del bot
def main():
    while True:
        try:
            print("Revisando balance...")
            balance = get_balance(wallet.public_key)
            print(f"Balance actual: {balance} SOL")
            if balance < BOT_BALANCE_THRESHOLD:
                print("El balance es insuficiente para continuar.")
                break

            print("Buscando tokens prometedores...")
            tokens = find_new_tokens()
            for token in tokens:
                if analyze_token(token):
                    print(f"Token prometedor encontrado: {token['name']}")
                    buy_token(token["address"], 1)  # Comprar 1 SOL de este token
                    time.sleep(5)  # Pausar para evitar spamming

            print("Esperando antes de la próxima iteración...")
            time.sleep(60)  # Esperar 1 minuto antes de repetir
        except Exception as e:
            print(f"Error en el bot: {e}")
            time.sleep(10)

if __name__ == "__main__":
    main()

