requirements.txt
solana==0.35.1
spl-token
requests
